Now take the jobs from the bag-of-tasks and register them in an atomical operation
This guarantees that no jobs are lost in case of a crash between the moment we remove the jobs from the bag of task and the moment they are inserted into the jobs table. Unfortunatelly, this shows the limits of datasets that we can't use in this case. Anyway, this makes the runner a bit simpler. 


git-svn-id: https://scm.gforge.inria.fr/svn/cigri/branches/cigri-3@926 a0e46a8e-710f-0410-a56b-8ab1cdea0c2d