Bugfix: http 400 errors correctly managed
400 error code (Bad Request) should be treated as a user error, and should not blacklist a cluster for everybody.
This fix adds support for 400 errors as "normal" campaign errors.
What this fix does not: 500 errors (Server Error) are treated as user errors but shouldn't. The problem is that the OAR API does not always return a 400 error for a bad request. Oarsub sometimes returns unknown error codes (like 251 for example) even for some bad requests (try to submit something with "-l dummyprop=1"). This situation should be clarified!